Add Seasons

Higher-order data summaries, by 6-month seasons, 3-month seasons, or year. Also applies specified functions to the specified summary level.

yearly generates a table of yearly means

add_seasons(
  data,
  level = "site",
  season_level = 2,
  date_column = "yearmon",
  summary_funs = NA,
  path = get_default_data_path(),
  download_if_missing = TRUE,
  clean = TRUE
)

yearly(...)

Arguments

data
data frame containing columns: date, period, newmoonnumber, or year and month

level
summarize by "Plot", "Treatment", or "Site"

season_level
either year, 2: winter = Oct-March summer = April-Sept 4: winter = Dec-Feb spring = March-May summer = Jun-Aug fall = Sep-Nov

date_column
either "date" (must be in format "y-m-d"), "period", "newmoonnumber", or "yearmon" (data must contain "year" and "month")

summary_funs
A function specified by its name (e.g. "mean"). Default is NA (returned with seasons added but not summarized).

path
either the file path that contains the PortalData folder or "repo", which then pulls data from the PortalData GitHub repository

download_if_missing
if the specified file path doesn't have the PortalData folder, then download it

clean
logical, load only QA/QC rodent data (TRUE) or all data (FALSE)

...
arguments passed to add_seasons

Value

a data.frame with additional "season" and "year" column, and other columns summarized as specified. If no summary function is specified, "season" and "year" columns are added to original dataframe, as well as a "seasonyear" column which correctly assigns months to seasons for grouping (eg December 2000 in winter 2001, rather than winter 2000).

Examples

# \donttest{
yearly(abundance(path = "repo", time = "newmoon"),
       date_column = "newmoonnumber", path = "repo")

#> Loading in data version 2.48.0

#> # A tibble: 41 x 22
#>     year    BA    DM    DO    DS  `NA`    OL    OT    PB    PE      PF     PH
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>   <dbl>  <dbl>
#>  1  1979     0  18.7  2.67 12.3   3.33  6.33 10.7      0  3.33  2.67   0     
#>  2  1980     0  22    4    21     3.92  5.92  5.67     0  1.33  6.83   0     
#>  3  1981     0  30.7  6.36 20.6   5.27  3.09  4.27     0  3.64  4.36   0     
#>  4  1982     0  47.3  9.2  28.5  10.1   5.3   6.6      0  9.6  12.7    0     
#>  5  1983     0  44   13.9  23.3   8.17  5.42  7.17     0  3.17  9.5    0     
#>  6  1984     0  32.3  8.2   6.4   5.9   6.4   4.5      0  1.5   0.6    0.7   
#>  7  1985     0  48.9 16.6   7.55  3.91  7.18  5.09     0  4.27  0.0909 0.273 
#>  8  1986     0  40.6 11.5   8.8   6     4.3   2.5      0  2.6   0.2    0.1   
#>  9  1987     0  42.5 19.4   9.45 11.6   7.18  3.91     0 12.9   0.364  0     
#> 10  1988     0  30.4 12     4.5   8.5   4.25  7.25     0 15.8   0.167  0.0833
#> # … with 31 more rows, and 10 more variables: PI <dbl>, PL <dbl>, PM <dbl>,
#> #   PP <dbl>, RF <dbl>, RM <dbl>, RO <dbl>, SF <dbl>, SH <dbl>, SO <dbl>

# }